Refactored signal.updates()
