+++ src/main/scala/com/raquo/airstream/core/Signal.scala
@@ -2,6 +2,7 @@ package com.raquo.airstream.core
 
 import com.raquo.airstream.combine.CombineSignalN
 import com.raquo.airstream.combine.generated.{CombineSignalObjectOps, CombineSignalOps}
+import com.raquo.airstream.conversions.{StreamFromSignal, UpdatesSignalOps}
 import com.raquo.airstream.core.Source.SignalSource
 import com.raquo.airstream.custom.{CustomSignalSource, CustomSource}
 import com.raquo.airstream.custom.CustomSource._
@@ -9,11 +10,12 @@ import com.raquo.airstream.debug.{Debugger, DebuggerSignal, DebugOps, DebugSigna
 import com.raquo.airstream.distinct.{DistinctOps, DistinctSignal}
 import com.raquo.airstream.dynamicImport.{DynamicImportSignalObjectOps, DynamicImportSignalOps}
 import com.raquo.airstream.extensions._
-import com.raquo.airstream.misc.{MapSignal, ScanLeftSignal, StreamFromSignal}
+import com.raquo.airstream.misc.{MapSignal, ScanLeftSignal}
 import com.raquo.airstream.ownership.Owner
 import com.raquo.airstream.state.{ObservedSignal, OwnedSignal, Val}
 import com.raquo.airstream.timing.JsPromiseSignal
 import com.raquo.ew.JsArray
+import org.scalajs.dom.svg.A
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.scalajs.js
@@ -27,6 +29,7 @@ with BaseObservable[Signal, A]
 with SignalSource[A]
 with CombineSignalOps[A] // combineWith, combineWithFn, withCurrentValueOf, sample
 with ScanLeftSignalOps[Signal, A] // scanLeft, scanLeftRecover
+with UpdatesSignalOps[A]
 with DebugSignalOps[Signal, A] // debug* (debugLogEvents, debugSpyAll, etc.)
 with DynamicImportSignalOps[A] // dynamicImport (Scala 3 only)
 {
@@ -67,70 +70,7 @@ with DynamicImportSignalOps[A] // dynamicImport (Scala 3 only)
     new MapSignal[A, B](parent = this, projectValue, recover = Some(recoverError))
   }
 
-  /** @param operator Note: Must not throw! */
-  def compose[B](operator: Signal[A] => Signal[B]): Signal[B] = {
-    operator(this)
-  }
-
-  /** Modify the Signal's changes stream, e.g. signal.composeChanges(_.delay(ms = 100))
-    *
-    * Alias to changes(operator). See also: [[composeAll]]
-    *
-    * @param operator Note: Must not throw!
-    */
-  def composeUpdates[AA >: A](
-    operator: EventStream[A] => EventStream[AA]
-  ): Signal[AA] = {
-    composeAll(updatesOperator = operator, initialOperator = identity)
-  }
-
-  /** Modify both the Signal's changes stream, and its initial.
-    * Similar to composeChanges, but lets you output a type unrelated to A.
-    *
-    * @param updatesOperator Note: Must not throw!
-    * @param initialOperator Note: Must not throw!
-    */
-  def composeAll[B](
-    updatesOperator: EventStream[A] => EventStream[B],
-    initialOperator: Try[A] => Try[B],
-    cacheInitialValue: Boolean = false
-  ): Signal[B] = {
-    updatesOperator(updates).toSignalWithTry(initialOperator(tryNow()), cacheInitialValue)
-  }
-
-  // #TODO[API] Why is .updates a def, and not a lazy val?
-  //  See `signal.updates shouldNotBe signal.updates` in SignalSpec
-  /** A stream of all values in this signal, excluding the initial value.
-    *
-    * When re-starting this stream, it emits the signal's new current value
-    * if and only if something has caused the signal's value to be updated
-    * or re-evaluated while the updates stream was stopped. This way the
-    * updates stream stays in sync with the signal even after restarting.
-    */
-  def updates: EventStream[A] = new StreamFromSignal[A](parent = this, updatesOnly = true)
-
-  /** Modify the Signal's updates, e.g. signal.updates(_.delay(ms = 100))
-    *
-    * Alias to [[composeUpdates]]. See also: [[composeAll]]
-    *
-    * @param compose Note: Must not throw!
-    */
-  @inline def updates[AA >: A](compose: EventStream[A] => EventStream[AA]): Signal[AA] = {
-    composeUpdates(compose)
-  }
-
-  @deprecated("signal.composeChanges renamed to signal.composeUpdates", since = "18.0.0-M3")
-  def composeChanges[AA >: A](
-    operator: EventStream[A] => EventStream[AA]
-  ): Signal[AA] = {
-    composeUpdates(operator)
-  }
-
-  @deprecated("signal.changes renamed to signal.updates", since = "18.0.0-M3")
-  def changes: EventStream[A] = updates
-
-  @deprecated("signal.changes renamed to signal.updates", since = "18.0.0-M3")
-  def changes[AA >: A](compose: EventStream[A] => EventStream[AA]): Signal[AA] = updates(compose)
+  @inline def updates: EventStream[A] = new StreamFromSignal[A](parent = this, updatesOnly = true)
 
   @deprecated("foldLeft was renamed to scanLeft", "15.0.0-M1")
   def foldLeft[B](makeInitial: A => B)(fn: (B, A) => B): Signal[B] = scanLeft(makeInitial)(fn)
